backend/requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17          # Required for UploadFile support
httpx==0.27.0                     # HTTP client for geocoding

# Database
motor==3.6.0
dnspython==2.7.0

# Config
pydantic-settings==2.6.1

# Image processing
Pillow==11.0.0
numpy==2.1.3

# AI Model (uncomment when using a real YOLOv8 model)
# ultralytics==8.3.0
# torch==2.5.1
# torchvision==0.20.1

faiss-cpu
sentence-transformers
